// packages/style/src/types.ts
export type Breakpoint = 'mobile' | 'tablet' | 'desktop' | 'wide';

export type StyleTarget = MaybeArray<CSSProp>;
export type StyleResolver<T = unknown> = (
  value: T
) => T | number | string | boolean | null | undefined;
// ...
export type ResponsiveProp<T> = Partial<Record<Breakpoint, T>>;
export type Responsive<T> = T | ResponsiveProp<T>;

type BorderStyle = 'none' | 'dashed' | 'dotted' | 'solid';
type OverflowValue = 'visible' | 'hidden' | 'scroll' | 'clip' | 'auto';
type BreakpointAbove = Exclude<Breakpoint, 'large'>;
type BreakpointBelow = Exclude<Breakpoint, 'base'>;
export type BreakpointRange =
  | { above: BreakpointAbove }
  | { below: BreakpointBelow }
  | Responsive<boolean>;